requirements.txt
streamlit
numpy<2
PyPDF2
sentence_transformers
oletools
reportlab
bcrypt
bandit
pylint
flake8
mypy
black
safety
requests
openai
python-dotenv
